Extract console package
