+++ src/main/scala/intellij/haskell/runconfig/console/HaskellConsoleConfiguration.scala
@@ -1,9 +1,10 @@
-package intellij.haskell.runconfig
+package intellij.haskell.runconfig.console
 
 import com.intellij.execution.Executor
 import com.intellij.execution.configurations._
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.openapi.project.Project
+import intellij.haskell.runconfig.HaskellStackConfigurationBase
 
 final class HaskellConsoleConfiguration(override val name: String, override val project: Project, override val configurationFactory: ConfigurationFactory)
   extends HaskellStackConfigurationBase(name, project, configurationFactory) {
