+++ src/main/scala/intellij/haskell/runconfig/console/HaskellConsoleConfigurationFactory.scala
@@ -1,4 +1,4 @@
-package intellij.haskell.runconfig
+package intellij.haskell.runconfig.console
 
 import com.intellij.execution.configurations.{ConfigurationFactory, ConfigurationType}
 import com.intellij.openapi.project.Project
