+++ src/main/scala/intellij/haskell/runconfig/console/HaskellConsoleView.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package intellij.haskell.runconfig
+package intellij.haskell.runconfig.console
 
 import java.io.{IOException, OutputStreamWriter}
