+++ src/main/scala/intellij/haskell/runconfig/console/HaskellConsoleViewDict.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package intellij.haskell.runconfig
+package intellij.haskell.runconfig.console
 
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
